feat(using-animated-api): Resets the example with an updated react-native version
