+++ example/.gitignore
@@ -1,41 +1,17 @@
-# OSX
-#
-.DS_Store
+# See https://help.github.com/ignore-files/ for more about ignoring files.
 
-# Xcode
-#
-build/
-*.pbxuser
-!default.pbxuser
-*.mode1v3
-!default.mode1v3
-*.mode2v3
-!default.mode2v3
-*.perspectivev3
-!default.perspectivev3
-xcuserdata
-*.xccheckout
-*.moved-aside
-DerivedData
-*.hmap
-*.ipa
-*.xcuserstate
-project.xcworkspace
+# expo
+.expo/
 
-# Android/IJ
-#
-*.iml
-.idea
-.gradle
-local.properties
+# dependencies
+/node_modules
 
-# node.js
-#
-node_modules/
-npm-debug.log
+# misc
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
 
-# BUCK
-buck-out/
-\.buckd/
-android/app/libs
-android/keystores/debug.keystore
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
