+++ example/app.json
@@ -1,0 +1,5 @@
+{
+  "expo": {
+    "sdkVersion": "23.0.0"
+  }
+}
